Change input of probesearch.py to accept alignment files
This should simplify the number of arguments that have to be passed to the program as you should only need to pass the alignment, and optionally a target start and end coordinate in the alignment.

Additionally the indexes should be fixed so that the probe generation starts at the first nucleotide position of the region, as there was no conversion from 1-based coordinates to 0-based coordinates before when passing the template sequence to the probeGenerator.

Finally, flags have been fixed. Will still have to remove a few more flags later because they're unnecessary.